--- README.md
@@ -1,12 +1,0 @@
-# 12th-fe-10th-assignment
-
-### 과제 소개
-- 무료로 사용할 수 있는 API를 활용하여 간단한 서비스를 제공하는 사이트를 만들어보세요!
-- 단순한 기능이라도 좋습니다. 직접 각 API의 docs를 찾아 읽어보고 이를 활용해 보는 것이 이번 과제의 주요 목표입니다!
-- 어떤 라이브러리를 사용하셔도 상관없습니다. 오히려 여러 react 관련 라이브러리를 사용해 보는 것도 좋은 경험이 될 것 같네요
-
-### 기한 : 6/2(PR 코드리뷰 포함)
-
-### 제출 방법
-- 이전 과제들과 동일하게 fork한 후 해당 repository에 과제를 진행하신 후 PR 넣어주시면 됩니다
-- 단, 내가 개발한 기능/유의 깊게 개발한 부분/개발할 때 들었던 의문사항을 상세하게 기재해 주세요!
