--- hw10/src/App.css
@@ -1,43 +1,0 @@
-
-.App {
-  text-align: center;
-  
-}
-
-
-.App-header {
-  background-color: #e50175;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.shocked{
-  width:300px;
-  height:200px;
-}
-.recom {
-  padding: 10px 20px;
-  font-size: 16px;
-  margin-top: 20px;
-  cursor: pointer;
-  border-color: white;
-  border-radius: 30px;
-  background-color:#e50175;
-  color: white;
-  font-weight: bold;
-  font-style: italic;
-  letter-spacing: 4px;
-}
-
-.recom:hover{
-  background-color:white;
-  color:#e50175;
-  transition: all 0.9s;
-  transform: rotate(360deg);
-}
-
