+++ hw10/src/Genre.css
@@ -1,0 +1,90 @@
+  @font-face {
+    font-family: 'gothic';
+    src: url('./font/gothic.ttf')format('truetype');
+  }
+  
+  @font-face {
+    font-family: 'gullim';
+    src: url('./font/gullim.TTF') format('truetype');
+  }
+
+  @font-face {
+    font-family: 'expo';
+    src: url('./font/expo.ttf');
+  }
+  
+  *{
+    background-color: #e50175;
+  }
+  
+  
+  .header {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: 50px;
+    color: white;
+  }
+  
+  div{
+    font-family: 'Gothic';
+    letter-spacing: 3px;
+  }
+  
+  .click{
+    font-family: 'gullim';
+    font-size:20px;
+    -webkit-text-stroke-width: 0.8px;
+     -webkit-text-stroke-color: white;
+     color:#e50175;
+     font-style: italic;
+  }
+  .content{
+    text-align: center ;
+    margin: 10px;
+    font-size: 30px;
+    font-family: sans-serif;
+    letter-spacing: 0;
+     border-color: white;
+  
+  }
+  
+  .content>p{
+    font-size: 15px;
+    font-family: 'gullim';
+    letter-spacing: 3px;
+  }
+  
+  .shocked{
+    width:300px;
+    height:200px;
+    margin: 10px;
+  }
+  .recom {
+    font-family: 'gullim';
+    padding: 10px 20px;
+    font-size: 20px;
+    margin: 10px;
+    cursor: pointer;
+    border-color: white;
+    border-radius: 30px;
+    background-color:#e50175;
+    color: white;
+    font-weight: bold;
+    font-style: italic;
+    letter-spacing: 4px;
+  }
+
+  .recom:hover{
+    background-color:white;
+    color:#e50175;
+    transition: all 0.9s;
+    transform: rotate(360deg);
+  }
+  
+  .bottom{
+    margin-top:20px;
+    letter-spacing: 15px;
+  }
